requirements.txt
beautifulsoup4==4.12.3
blinker==1.9.0
click==8.1.8
colorama==0.4.6
EbookLib==0.18
EbookLib-MD==0.22
Flask==3.1.0
itsdangerous==2.2.0
Jinja2==3.1.5
lxml==5.3.0
MarkupSafe==3.0.2
pillow==11.1.0
six==1.17.0
soupsieve==2.6
Werkzeug==3.1.3
